requirements.txt
agno
python-dotenv
openai
duckduckgo-search
fastapi
uvicorn
pydantic[dotenv]
pyyaml
ddgs
sqlalchemy
